memory

+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/dataobject/MemoryCardDO.java
@@ -15,7 +15,7 @@ public class MemoryCardDO {
 
     private Long currentVersionId;
 
-    private Integer state;
+    private Byte state;
 
     private LocalDateTime createdAt;
